rename json-schema to JSONSchema
