+++ catalog/app/components/JsonEditor/State.ts
@@ -4,8 +4,8 @@ import * as R from 'ramda'
 import * as React from 'react'
 
 import * as JSONPointer from 'utils/JSONPointer'
-import { JsonSchema } from 'utils/json-schema'
-import * as jsonSchemaUtils from 'utils/json-schema/json-schema'
+import type { JsonSchema } from 'utils/JSONSchema'
+import * as JSONSchema from 'utils/JSONSchema'
 import { Json, JsonRecord } from 'utils/types'
 
 import {
@@ -77,7 +77,7 @@ type SortOrder = React.MutableRefObject<{
   dict: Record<JSONPointer.Pointer, number>
 }>
 
-// TODO: consider to use 'json-schema-traverse'
+// TODO: consider to use 'JSONSchema-traverse'
 // NOTE: memo is mutated, sortOrder is React.ref and mutated too
 export function iterateSchema(
   schema: JsonSchema,
@@ -162,10 +162,10 @@ function calcReactId(valuePath: JSONPointer.Path, value?: Json): string {
 function getDefaultValue(jsonDictItem?: SchemaItem): Json | typeof EMPTY_VALUE {
   if (!jsonDictItem?.valueSchema) return EMPTY_VALUE
 
-  const defaultFromSchema = jsonSchemaUtils.getDefaultValue(jsonDictItem?.valueSchema)
+  const defaultFromSchema = JSONSchema.getDefaultValue(jsonDictItem?.valueSchema)
   if (defaultFromSchema !== undefined) return defaultFromSchema
 
-  const setDefaults = jsonSchemaUtils.makeSchemaDefaultsSetter(jsonDictItem?.valueSchema)
+  const setDefaults = JSONSchema.makeSchemaDefaultsSetter(jsonDictItem?.valueSchema)
   const nestedDefaultFromSchema = setDefaults(undefined)
   if (nestedDefaultFromSchema !== undefined) return nestedDefaultFromSchema
