+++ catalog/app/containers/Bucket/PackageDialog/MetaInput.tsx
@@ -13,7 +13,7 @@ import JsonValidationErrors from 'components/JsonValidationErrors'
 import MetadataEditor from 'components/MetadataEditor'
 import * as Notifications from 'containers/Notifications'
 import useDragging from 'utils/dragging'
-import { JsonSchema } from 'utils/json-schema'
+import type { JsonSchema } from 'utils/JSONSchema'
 import * as spreadsheets from 'utils/spreadsheets'
 import { readableBytes } from 'utils/string'
 import { JsonRecord } from 'utils/types'
