+++ catalog/app/containers/Bucket/Upload.tsx
@@ -7,7 +7,7 @@ import type * as RF from 'react-final-form'
 import { JsonValue } from 'components/JsonEditor/constants'
 import useDragging from 'utils/dragging'
 import * as IPC from 'utils/electron/ipc-provider'
-import { JsonSchema } from 'utils/json-schema'
+import { JsonSchema } from 'utils/JSONSchema'
 import * as workflows from 'utils/workflows'
 import { getMetaValue, getWorkflowApiParam } from 'containers/Bucket/requests/package'
