+++ catalog/app/containers/Bucket/requests/requestsUntyped.js
@@ -9,7 +9,7 @@ import quiltSummarizeSchema from 'schemas/quilt_summarize.json'
 import { SUPPORTED_EXTENSIONS as IMG_EXTS } from 'components/Thumbnail'
 import * as quiltConfigs from 'constants/quiltConfigs'
 import * as Resource from 'utils/Resource'
-import { makeSchemaValidator } from 'utils/json-schema'
+import { makeSchemaValidator } from 'utils/JSONSchema'
 import mkSearch from 'utils/mkSearch'
 import * as s3paths from 'utils/s3paths'
 import tagged from 'utils/tagged'
