--- catalog/app/utils/json-schema/index.js
@@ -1,1 +1,0 @@
-export * from './json-schema'
